Split benchmark suite used for attention operators
